No more NodeRefs, builtin symbols, math, signed numbers
Nodes are no longer NodeRefs. Anything that contains data
either already contains a reference of some kind or holds
owned data that can be cloned. Is it the most effecient?
Probably not, but it's certainly easier to start with.

Builtin functions are now part of the normal symbol table
and symbols are now correctly evaulated to a function.

Added the four main math functions.

Made number objects signed.