Keywords, equality, logic, truthiness
Keywords are special symbols that always represent
themselves. They cannot have values or be called
as functions. They begin with a colon ":thing".

Implemented "eq", "not", "and", and "or".

Made truthiness consistant and simplier. Booleans are
represented with the keywords ":t" and ":f" for true
and false respectivly. "nil" is represented as ":f"
as many lisps don't differentiate between nil and false.
Maybe this does? Maybe add ":nil" as 'special'? Who knows.