Fixed bug where redefining a running function paniced
If function ran that redefined itself, the program would crash
with a borrow error. The symbol was being borrowed at eval and
then tried to be borrowed mutable when setf was trying to run.
This caused a panic since a mutable borrow can't be taken
at the same time as an immutable one.

The eval function now clones the symbol before executing
the function so there's no borrow conflict.